requirements.txt
python-dotenv
pandas
ipython
telnyx
certifi
pyopenssl
flask
python-dotenv
h5py
python-levenshtein
tables
pyflakes
psutil
redis
apscheduler
flask-apscheduler
pipreqs
bs4
lxml
nltk
numpy
redis
multiprocess
ipython-reload
panphon
redis-shelve
phonemizer
py-espeak-ng
hiredis
word2number
